+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # PyUbee CHANGELOG
 This file is used to list changes made in each version of the PyUbee.
 
+## 0.3 (unreleased)
+* Add pyubee command line interface
+* Add support for EVW3200-Wifi model
+
 ## 0.2 (January 13 2019)
 * Get list of WiFi and LAN devices instead of WiFi devices only
 * Updated README.md and CHANGELOG.md